third_party/blink/web_tests/external/wpt/css/css-scoping/keyframes-002-expected.txt

This is a testharness.js-based test.
[FAIL] @keyframes from the document should apply in the shadow tree
  assert_equals: expected 1 but got 0
Harness: the test ran to completion.
